CryptoUtils: move remaining HDUtils methods and remove HDUtils class
* Move the 3 remaining static methods from HDUtils to CryptoUtils
* Remove HDUtils
* Move HDUtilsTest to internal/CryptoUtilsTest
