--- core/src/main/java/org/bitcoinj/crypto/HDUtils.java
@@ -1,47 +1,0 @@
-/*
- * Copyright 2013 Matija Mazi.
- * Copyright 2014 Giannis Dzegoutanis.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.bitcoinj.crypto;
-
-import org.bouncycastle.crypto.digests.SHA512Digest;
-import org.bouncycastle.crypto.macs.HMac;
-import org.bouncycastle.crypto.params.KeyParameter;
-
-/**
- * Static utilities used in BIP 32 Hierarchical Deterministic Wallets (HDW).
- */
-public final class HDUtils {
-
-    static HMac createHmacSha512Digest(byte[] key) {
-        SHA512Digest digest = new SHA512Digest();
-        HMac hMac = new HMac(digest);
-        hMac.init(new KeyParameter(key));
-        return hMac;
-    }
-
-    static byte[] hmacSha512(HMac hmacSha512, byte[] input) {
-        hmacSha512.reset();
-        hmacSha512.update(input, 0, input.length);
-        byte[] out = new byte[64];
-        hmacSha512.doFinal(out, 0);
-        return out;
-    }
-
-    public static byte[] hmacSha512(byte[] key, byte[] data) {
-        return hmacSha512(createHmacSha512Digest(key), data);
-    }
-}
