+++ core/src/main/java/org/bitcoinj/crypto/internal/CryptoUtils.java
@@ -17,6 +17,9 @@ package org.bitcoinj.crypto.internal;
 
 import org.bitcoinj.base.Sha256Hash;
 import org.bouncycastle.crypto.digests.RIPEMD160Digest;
+import org.bouncycastle.crypto.digests.SHA512Digest;
+import org.bouncycastle.crypto.macs.HMac;
+import org.bouncycastle.crypto.params.KeyParameter;
 import org.bouncycastle.jcajce.provider.digest.SHA3;
 
 import java.nio.charset.StandardCharsets;
@@ -47,5 +50,24 @@ public class CryptoUtils {
         byte[] ripmemdHash = new byte[20];
         digest.doFinal(ripmemdHash, 0);
         return ripmemdHash;
+    }
+
+    public static HMac createHmacSha512Digest(byte[] key) {
+        SHA512Digest digest = new SHA512Digest();
+        HMac hMac = new HMac(digest);
+        hMac.init(new KeyParameter(key));
+        return hMac;
+    }
+
+    public static byte[] hmacSha512(HMac hmacSha512, byte[] input) {
+        hmacSha512.reset();
+        hmacSha512.update(input, 0, input.length);
+        byte[] out = new byte[64];
+        hmacSha512.doFinal(out, 0);
+        return out;
+    }
+
+    public static byte[] hmacSha512(byte[] key, byte[] data) {
+        return hmacSha512(createHmacSha512Digest(key), data);
     }
 }
